pyproject.toml
[project]
name = "log-analyzer-imp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.1",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.33",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "regex>=2024.11.6",
    "sqlalchemy>=2.0.25",
    "streamlit>=1.39.0",
]
